// src/types/stakingStats.ts
export interface GlobalStats {
    stakerCount: string
    activeStaked: string
    today: string
    yesterday: string
    thisWeek: string
    lastWeek: string
    thisMonth: string
    lastMonth: string
    totalRewards: string
    contractUsdt: string
    insurancePool: string
    totalClaimed: string
    feeToPool: string
    feeToInsurance: string
}

export interface WalletStats {
    activeStaked: string
    myShare: string
    pendingUnstake: string
    unlockTime: string
    withdrawableRewards: string
    totalEarnings: string
}